Move networking classes to main package
