+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/wooshippinglabels/packages/datasource/WooShippingLabelPackageRepository.kt
@@ -1,8 +1,8 @@
 package com.woocommerce.android.ui.orders.wooshippinglabels.packages.datasource
 
 import com.woocommerce.android.tools.SelectedSite
-import com.woocommerce.android.ui.orders.wooshippinglabels.packages.datasource.networking.PackageResponse
-import com.woocommerce.android.ui.orders.wooshippinglabels.packages.datasource.networking.WooShippingLabelPackageRestClient
+import com.woocommerce.android.ui.orders.wooshippinglabels.packages.networking.PackageResponse
+import com.woocommerce.android.ui.orders.wooshippinglabels.packages.networking.WooShippingLabelPackageRestClient
 import javax.inject.Inject
 import javax.inject.Singleton
 import org.wordpress.android.fluxc.model.SiteModel
